feat: use pages to separate the different part of the tool
